Extract waveform widget
