+++ src/gui_app/waveform.rs
@@ -1,28 +1,13 @@
 #![allow(missing_docs)]
 
-use radiant::{
-    gui::types::{Point, Rect, Vector2},
-    gui::visualization::TimelineEditPreview,
-    layout::LayoutOutput,
-    prelude as ui,
-    runtime::PaintPrimitive,
-    theme::ThemeTokens,
-    widgets::{
-        FocusBehavior, PaintBounds, PointerButton, Widget, WidgetCommon, WidgetInput, WidgetOutput,
-        WidgetSizing,
-    },
-};
+use radiant::prelude as ui;
 use std::{path::PathBuf, sync::Arc};
-
-use super::GuiMessage;
 
 const WAVEFORM_WIDTH: usize = 1200;
 const WAVEFORM_HEIGHT: usize = 320;
 const MIN_VISIBLE_FRAMES: usize = 256;
 const BAND_COUNT: usize = 4;
 const SELECTION_DRAG_EPSILON: f32 = 0.001;
-const SELECTION_MOVE_HANDLE_HEIGHT: f32 = 7.0;
-const SELECTION_MOVE_HANDLE_END_INSET: f32 = 9.0;
 const SELECTION_FLASH_FRAMES: u8 = 12;
 #[cfg(test)]
 const SYNTHETIC_SAMPLE_RATE: u32 = 48_000;
@@ -264,442 +249,17 @@ use audio_file::{
 mod signal_widget;
 use signal_widget::WaveformSignalWidget;
 
+mod widget;
+#[cfg(test)]
+pub(super) use widget::WaveformWidgetProps;
+pub(super) use widget::{WaveformWidget, waveform_viewport_view};
+
+mod widget_geometry;
+
 mod edit_fade_paint;
 mod selection_paint;
 
 pub(super) type WaveformViewport = ui::IndexViewport;
-
-pub(super) fn waveform_viewport_view(state: &WaveformState) -> ui::View<super::GuiMessage> {
-    ui::stack([
-        ui::custom_widget(
-            WaveformSignalWidget::new(
-                state.file(),
-                state.viewport(),
-                state.edit_selection(),
-                state.active_drag_kind(),
-            ),
-            |_| None,
-        )
-        .id(11)
-        .size(WAVEFORM_WIDTH as f32, WAVEFORM_HEIGHT as f32),
-        ui::custom_widget(
-            WaveformWidget::new(WaveformWidgetProps::from_state(state)),
-            |output| {
-                output
-                    .typed_ref::<WaveformInteraction>()
-                    .copied()
-                    .map(GuiMessage::Waveform)
-            },
-        )
-        .id(12)
-        .size(WAVEFORM_WIDTH as f32, WAVEFORM_HEIGHT as f32),
-    ])
-    .id(10)
-    .size(WAVEFORM_WIDTH as f32, WAVEFORM_HEIGHT as f32)
-}
-
-#[derive(Clone, Debug)]
-struct WaveformWidgetProps {
-    file: Arc<WaveformFile>,
-    viewport: WaveformViewport,
-    playhead_ratio: Option<f32>,
-    play_mark_ratio: Option<f32>,
-    edit_mark_ratio: Option<f32>,
-    play_selection: Option<wavecrate::selection::SelectionRange>,
-    edit_selection: Option<wavecrate::selection::SelectionRange>,
-    play_selection_flash_frames: u8,
-    active_drag_kind: Option<WaveformActiveDragKind>,
-}
-
-impl WaveformWidgetProps {
-    fn from_state(state: &WaveformState) -> Self {
-        Self {
-            file: state.file(),
-            viewport: state.viewport(),
-            playhead_ratio: state.playhead_ratio(),
-            play_mark_ratio: state.play_mark_ratio(),
-            edit_mark_ratio: state.edit_mark_ratio(),
-            play_selection: state.play_selection(),
-            edit_selection: state.edit_selection(),
-            play_selection_flash_frames: state.play_selection_flash_frames(),
-            active_drag_kind: state.active_drag_kind(),
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-struct WaveformWidget {
-    common: WidgetCommon,
-    file: Arc<WaveformFile>,
-    viewport: WaveformViewport,
-    playhead_ratio: Option<f32>,
-    play_mark_ratio: Option<f32>,
-    edit_mark_ratio: Option<f32>,
-    play_selection: Option<wavecrate::selection::SelectionRange>,
-    edit_selection: Option<wavecrate::selection::SelectionRange>,
-    play_selection_flash_frames: u8,
-    edit_preview: TimelineEditPreview,
-    active_drag_kind: Option<WaveformActiveDragKind>,
-}
-
-impl WaveformWidget {
-    fn new(props: WaveformWidgetProps) -> Self {
-        let WaveformWidgetProps {
-            file,
-            viewport,
-            playhead_ratio,
-            play_mark_ratio,
-            edit_mark_ratio,
-            play_selection,
-            edit_selection,
-            play_selection_flash_frames,
-            active_drag_kind,
-        } = props;
-        let mut common = WidgetCommon::new(
-            0,
-            WidgetSizing::fixed(Vector2::new(WAVEFORM_WIDTH as f32, WAVEFORM_HEIGHT as f32)),
-        );
-        common.focus = FocusBehavior::Pointer;
-        common.paint.bounds = PaintBounds::ClipToRect;
-        common.paint.paints_focus = false;
-        common.paint.paints_state_layers = false;
-        Self {
-            common,
-            file,
-            viewport,
-            playhead_ratio,
-            play_mark_ratio,
-            edit_mark_ratio,
-            play_selection,
-            edit_selection,
-            play_selection_flash_frames,
-            edit_preview: edit_preview_for_selection(edit_selection),
-            active_drag_kind,
-        }
-    }
-
-    fn ratio_from_position(&self, bounds: Rect, position: Point) -> f32 {
-        ((position.x - bounds.min.x) / bounds.width().max(1.0)).clamp(0.0, 1.0)
-    }
-}
-
-impl Widget for WaveformWidget {
-    fn common(&self) -> &WidgetCommon {
-        &self.common
-    }
-
-    fn common_mut(&mut self) -> &mut WidgetCommon {
-        &mut self.common
-    }
-
-    fn handle_input(&mut self, bounds: Rect, input: WidgetInput) -> Option<WidgetOutput> {
-        match input {
-            WidgetInput::PointerMove { position } => {
-                self.common.state.hovered = bounds.contains(position);
-                self.active_drag_kind.map(|_| {
-                    WidgetOutput::typed(WaveformInteraction::UpdateSelection {
-                        visible_ratio: self.ratio_from_position(bounds, position),
-                    })
-                })
-            }
-            WidgetInput::Wheel { position, delta } if bounds.contains(position) => {
-                Some(WidgetOutput::typed(WaveformInteraction::Wheel {
-                    delta,
-                    anchor_ratio: self.ratio_from_position(bounds, position),
-                }))
-            }
-            WidgetInput::PointerPress {
-                position,
-                button: PointerButton::Primary,
-                ..
-            } if bounds.contains(position) => {
-                if let Some(handle) = self.edit_fade_handle_at(bounds, position) {
-                    return Some(WidgetOutput::typed(WaveformInteraction::BeginEditFade {
-                        handle,
-                        visible_ratio: self.ratio_from_position(bounds, position),
-                    }));
-                }
-                if let Some(edge) =
-                    self.selection_resize_handle_at(bounds, position, WaveformSelectionKind::Play)
-                {
-                    return Some(WidgetOutput::typed(
-                        WaveformInteraction::BeginSelectionResize {
-                            kind: WaveformSelectionKind::Play,
-                            edge,
-                            visible_ratio: self.ratio_from_position(bounds, position),
-                        },
-                    ));
-                }
-                if self.selection_move_handle_at(bounds, position, WaveformSelectionKind::Play) {
-                    return Some(WidgetOutput::typed(
-                        WaveformInteraction::BeginSelectionMove {
-                            kind: WaveformSelectionKind::Play,
-                            visible_ratio: self.ratio_from_position(bounds, position),
-                        },
-                    ));
-                }
-                if self.selection_move_handle_at(bounds, position, WaveformSelectionKind::Edit) {
-                    return Some(WidgetOutput::typed(
-                        WaveformInteraction::BeginSelectionMove {
-                            kind: WaveformSelectionKind::Edit,
-                            visible_ratio: self.ratio_from_position(bounds, position),
-                        },
-                    ));
-                }
-                Some(WidgetOutput::typed(WaveformInteraction::BeginSelection {
-                    kind: WaveformSelectionKind::Play,
-                    visible_ratio: self.ratio_from_position(bounds, position),
-                }))
-            }
-            WidgetInput::PointerDoubleClick {
-                position,
-                button: PointerButton::Primary,
-                ..
-            } if bounds.contains(position) => {
-                if let Some(
-                    handle @ (WaveformEditFadeHandle::FadeInOuterStart
-                    | WaveformEditFadeHandle::FadeOutOuterEnd),
-                ) = self.edit_fade_handle_at(bounds, position)
-                {
-                    return Some(WidgetOutput::typed(
-                        WaveformInteraction::ClearEditFadeSilence { handle },
-                    ));
-                }
-                None
-            }
-            WidgetInput::PointerPress {
-                position,
-                button: PointerButton::Secondary,
-                ..
-            } if bounds.contains(position) => {
-                if let Some(handle) = self.edit_fade_handle_at(bounds, position) {
-                    return Some(WidgetOutput::typed(WaveformInteraction::BeginEditFade {
-                        handle,
-                        visible_ratio: self.ratio_from_position(bounds, position),
-                    }));
-                }
-                if self.selection_move_handle_at(bounds, position, WaveformSelectionKind::Edit) {
-                    return Some(WidgetOutput::typed(
-                        WaveformInteraction::BeginSelectionMove {
-                            kind: WaveformSelectionKind::Edit,
-                            visible_ratio: self.ratio_from_position(bounds, position),
-                        },
-                    ));
-                }
-                Some(WidgetOutput::typed(WaveformInteraction::BeginSelection {
-                    kind: WaveformSelectionKind::Edit,
-                    visible_ratio: self.ratio_from_position(bounds, position),
-                }))
-            }
-            WidgetInput::PointerPress {
-                position,
-                button: PointerButton::Auxiliary,
-                ..
-            } if bounds.contains(position) => {
-                Some(WidgetOutput::typed(WaveformInteraction::BeginPan {
-                    visible_ratio: self.ratio_from_position(bounds, position),
-                }))
-            }
-            WidgetInput::PointerRelease {
-                position,
-                button: PointerButton::Primary,
-                ..
-            } if self.active_drag_kind
-                == Some(WaveformActiveDragKind::Selection(
-                    WaveformSelectionKind::Play,
-                ))
-                || matches!(
-                    self.active_drag_kind,
-                    Some(
-                        WaveformActiveDragKind::EditFade(_)
-                            | WaveformActiveDragKind::SelectionResize(
-                                WaveformSelectionKind::Play,
-                                _
-                            )
-                            | WaveformActiveDragKind::SelectionMove(_)
-                    )
-                ) =>
-            {
-                Some(WidgetOutput::typed(WaveformInteraction::FinishSelection {
-                    visible_ratio: self.ratio_from_position(bounds, position),
-                }))
-            }
-            WidgetInput::PointerRelease {
-                position,
-                button: PointerButton::Secondary,
-                ..
-            } if self.active_drag_kind
-                == Some(WaveformActiveDragKind::Selection(
-                    WaveformSelectionKind::Edit,
-                ))
-                || matches!(
-                    self.active_drag_kind,
-                    Some(
-                        WaveformActiveDragKind::EditFade(_)
-                            | WaveformActiveDragKind::SelectionMove(WaveformSelectionKind::Edit)
-                    )
-                ) =>
-            {
-                Some(WidgetOutput::typed(WaveformInteraction::FinishSelection {
-                    visible_ratio: self.ratio_from_position(bounds, position),
-                }))
-            }
-            WidgetInput::PointerRelease {
-                position,
-                button: PointerButton::Auxiliary,
-                ..
-            } if self.active_drag_kind == Some(WaveformActiveDragKind::Pan) => {
-                Some(WidgetOutput::typed(WaveformInteraction::FinishSelection {
-                    visible_ratio: self.ratio_from_position(bounds, position),
-                }))
-            }
-            _ => None,
-        }
-    }
-
-    fn accepts_wheel_input(&self) -> bool {
-        true
-    }
-
-    fn append_paint(
-        &self,
-        primitives: &mut Vec<PaintPrimitive>,
-        bounds: Rect,
-        _layout: &LayoutOutput,
-        _theme: &ThemeTokens,
-    ) {
-        self.append_selection_and_marker_paint(primitives, bounds);
-        self.append_edit_fade_paint(primitives, bounds);
-    }
-}
-
-impl WaveformWidget {
-    fn selection_move_handle_at(
-        &self,
-        bounds: Rect,
-        position: Point,
-        kind: WaveformSelectionKind,
-    ) -> bool {
-        let range = match kind {
-            WaveformSelectionKind::Play => self.play_selection,
-            WaveformSelectionKind::Edit => self.edit_selection,
-        };
-        let Some((start, end)) = self.visible_range_for_selection(range) else {
-            return false;
-        };
-        self.selection_move_handle_rect(bounds, start, end)
-            .is_some_and(|rect| rect.contains(position))
-    }
-
-    fn selection_move_handle_rect(&self, bounds: Rect, start: f32, end: f32) -> Option<Rect> {
-        let left = bounds.min.x + bounds.width() * start.min(end).clamp(0.0, 1.0);
-        let right = bounds.min.x + bounds.width() * start.max(end).clamp(0.0, 1.0);
-        if right <= left {
-            return None;
-        }
-        let width = right - left;
-        let inset = SELECTION_MOVE_HANDLE_END_INSET.min(width * 0.28);
-        let handle_left = if width > inset * 2.0 + 1.0 {
-            left + inset
-        } else {
-            left
-        };
-        let handle_right = if width > inset * 2.0 + 1.0 {
-            right - inset
-        } else {
-            right
-        };
-        let height = SELECTION_MOVE_HANDLE_HEIGHT
-            .min(bounds.height().max(1.0))
-            .max(1.0);
-        let handle_right = handle_right.max(handle_left + 1.0).min(bounds.max.x);
-        if handle_right <= handle_left {
-            return None;
-        }
-        Some(Rect::from_min_max(
-            Point::new(handle_left, bounds.min.y),
-            Point::new(handle_right, bounds.min.y + height),
-        ))
-    }
-
-    fn selection_resize_handle_at(
-        &self,
-        bounds: Rect,
-        position: Point,
-        kind: WaveformSelectionKind,
-    ) -> Option<WaveformSelectionEdge> {
-        let range = match kind {
-            WaveformSelectionKind::Play => self.play_selection,
-            WaveformSelectionKind::Edit => self.edit_selection,
-        };
-        let (start, end) = self.visible_range_for_selection(range)?;
-        [WaveformSelectionEdge::Start, WaveformSelectionEdge::End]
-            .into_iter()
-            .find(|edge| {
-                self.selection_resize_handle_rect(bounds, start, end, *edge)
-                    .is_some_and(|rect| rect.contains(position))
-            })
-    }
-
-    fn selection_resize_handle_rect(
-        &self,
-        bounds: Rect,
-        start: f32,
-        end: f32,
-        edge: WaveformSelectionEdge,
-    ) -> Option<Rect> {
-        let x_ratio = match edge {
-            WaveformSelectionEdge::Start => start,
-            WaveformSelectionEdge::End => end,
-        };
-        let x = bounds.min.x + bounds.width() * x_ratio.clamp(0.0, 1.0);
-        let width = 7.0_f32.min(bounds.width().max(1.0));
-        let half_width = width * 0.5;
-        let top = bounds.min.y;
-        let bottom = (bounds.min.y + 22.0)
-            .min(bounds.max.y)
-            .max(bounds.min.y + 1.0);
-        let left = (x - half_width).clamp(bounds.min.x, bounds.max.x - width.max(1.0));
-        let right = (left + width).min(bounds.max.x).max(left + 1.0);
-        Some(Rect::from_min_max(
-            Point::new(left, top),
-            Point::new(right, bottom),
-        ))
-    }
-
-    fn visible_range_for_selection(
-        &self,
-        range: Option<wavecrate::selection::SelectionRange>,
-    ) -> Option<(f32, f32)> {
-        let range = range?;
-        let total = self.file.frames.max(1) as f32;
-        let visible_start = self.viewport.start as f32;
-        let visible_end = self.viewport.end as f32;
-        let visible_width = self.viewport.visible_items() as f32;
-        let start_frame = range.start().clamp(0.0, 1.0) * total;
-        let end_frame = range.end().clamp(0.0, 1.0) * total;
-        let left_frame = start_frame.min(end_frame).max(visible_start);
-        let right_frame = start_frame.max(end_frame).min(visible_end);
-        if right_frame <= left_frame {
-            return None;
-        }
-        let start = ((left_frame - visible_start) / visible_width.max(1.0)).clamp(0.0, 1.0);
-        let end = ((right_frame - visible_start) / visible_width.max(1.0)).clamp(0.0, 1.0);
-        Some((start, end))
-    }
-
-    fn visible_ratio_for_absolute(&self, ratio: Option<f32>) -> Option<f32> {
-        let absolute_ratio = ratio?;
-        let frame = absolute_ratio.clamp(0.0, 1.0) * self.file.frames.max(1) as f32;
-        let visible_start = self.viewport.start as f32;
-        let visible_width = self.viewport.visible_items() as f32;
-        let visible_ratio = (frame - visible_start) / visible_width.max(1.0);
-        if !(0.0..=1.0).contains(&visible_ratio) {
-            return None;
-        }
-        Some(visible_ratio)
-    }
-}
 
 #[cfg(test)]
 mod tests {
